replaced the arg type that specifies the length of bcmp() and bzero() from int to size_t (int is 4 bytes on both 32bit and 64bit archs, long, size_t are 4 bytes on 32bit and 8 bytes on 64bit)
git-svn-id: https://svn.aros.org/svn/aros/trunk@23133 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ compiler/clib/include/string.h
@@ -78,9 +78,9 @@ int stcu_d(char * out, unsigned uivalue); /* SAS C */
 void * memccpy (void * restrict dest, const void * restrict src,
 		    int c, size_t n);
 
-int bcmp (const void * s1, const void * s2, int n);
+int bcmp (const void * s1, const void * s2, size_t n);
 void bcopy (const void * src, void * dst, size_t n);
-void bzero (void * src, int n);
+void bzero (void * src, size_t n);
 
 char *strsep(char **, const char *); /* BSD */
